+++ back-end/controllers/songController.js
@@ -1,8 +1,8 @@
 const express = require("express")
 const song = express.Router()
-const { getAllSongs, getASong } = require('../queries/playlists')
+const { getAllSongs, getASong, createSong, deleteSong } = require('../queries/songs')
 
-
+//index route
 song.get("/", async (req, res) => {
    const allSongs = await getAllSongs();
 
@@ -13,6 +13,7 @@ song.get("/", async (req, res) => {
    }
 })
 
+//show route
 song.get('/:id', async (req, res)=>{
     const { id } = req.params;
     const song = await getASong(id);
@@ -23,5 +24,34 @@ song.get('/:id', async (req, res)=>{
         res.status(500).json({error: 'Server Error'})
     }
 })
+
+//create route
+song.post("/", async (req, res) =>{
+    const newSong = req.body
+    try {
+        const addedSong = await createSong(newSong)
+        res.status(202).json(addedSong)
+    }catch (error) {
+        res.status(400).json({error: error})
+    }
+})
+
+//delete route
+
+song.delete("/:id", async (req, res) =>{
+    const {id}  = req.params;
+    
+    try {
+        const deletedSong = await deleteSong(id);
+        res.status(200).json(deletedSong)
+
+    } catch (error) {
+        res.status(400).json({error: error})
+    }
+})
+
+
+
+
 
 module.exports = song;
