+++ back-end/db/seed.sql
@@ -1,6 +1,6 @@
-\c playlists_dev;
+-- \c playlists_dev;
 
-INSERT INTO bookmarks (name, artist, album, time, is_favorite) VALUES
-('', '', '', '', true),
-('', '', '', '', true),
-('', '', '', '', true);
+-- INSERT INTO bookmarks (name, artist, album, time, is_favorite) VALUES
+-- ('', '', '', '', true),
+-- ('', '', '', '', true),
+-- ('', '', '', '', true);
