--- back-end/queries/playlists.js
@@ -1,27 +1,0 @@
-const db = require ('../db/dbConfig');
-
-const getAllSongs = async () => {
-    try {
-        const allSongs = await db.any("SELECT * FROM songs");
-        return allSongs;
-
-    } catch (error){
-        return error;
-    }
-};
-
-const getASong = async (id) =>{
-    try {
-
-        const Song = await db.one("SELECT * FROM songs WHERE id=$1", id)
-        return song
-
-    }catch (error){
-        return error;
-    }
-};
-
-module.exports = {
-    getAllSongs,
-    getASong
-}
